Refactored. Protected from conflicting deletes.

+++ src/main/java/pl/tul/zzpj/dietmaster/logic/services/implementations/NutrientServiceImpl.java
@@ -6,13 +6,15 @@ import org.springframework.stereotype.Service;
 import pl.tul.zzpj.dietmaster.logic.controllers.requests.nutrient.CreateNutrientDto;
 import pl.tul.zzpj.dietmaster.logic.controllers.requests.nutrient.GetNutrientDto;
 import pl.tul.zzpj.dietmaster.logic.controllers.requests.nutrient.UpdateNutrientDto;
+import pl.tul.zzpj.dietmaster.logic.repositories.IngredientNutritionRepository;
 import pl.tul.zzpj.dietmaster.logic.repositories.NutrientRepository;
 import pl.tul.zzpj.dietmaster.logic.services.interfaces.NutrientService;
 import pl.tul.zzpj.dietmaster.model.entities.Nutrient;
 import pl.tul.zzpj.dietmaster.model.entities.enums.categories.NutrientCategory;
 import pl.tul.zzpj.dietmaster.model.exception.exists.NutrientExistsException;
+import pl.tul.zzpj.dietmaster.model.exception.notfound.NutrientNotFoundException;
+import pl.tul.zzpj.dietmaster.model.exception.used.NutrientUsedInIngredient;
 
-import javax.ws.rs.NotFoundException;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -26,10 +28,17 @@ public class NutrientServiceImpl implements NutrientService {
 
     private final NutrientRepository repository;
 
+    private final IngredientNutritionRepository ingredientRepository;
+
     @Autowired
-    public NutrientServiceImpl(NutrientRepository repository, ModelMapper mapper) {
+    public NutrientServiceImpl(
+        NutrientRepository repository,
+        ModelMapper mapper,
+        IngredientNutritionRepository ingredientRepository
+    ) {
         this.repository = repository;
         this.modelMapper = mapper;
+        this.ingredientRepository = ingredientRepository;
     }
 
     @Override
@@ -51,7 +60,8 @@ public class NutrientServiceImpl implements NutrientService {
     }
 
     @Override
-    public void updateNutrient(UpdateNutrientDto updateNutrientDto) throws NutrientExistsException {
+    public void updateNutrient(UpdateNutrientDto updateNutrientDto)
+        throws NutrientExistsException, NutrientNotFoundException {
 
         String newName = updateNutrientDto.getName();
         Long id = updateNutrientDto.getId();
@@ -64,7 +74,9 @@ public class NutrientServiceImpl implements NutrientService {
     }
 
     @Override
-    public void createNutrient(CreateNutrientDto createNutrientDto) throws NutrientExistsException {
+    public void createNutrient(CreateNutrientDto createNutrientDto)
+        throws NutrientExistsException {
+
         checkIfNameExists(createNutrientDto.getName());
 
         Nutrient nutrient = modelMapper.map(createNutrientDto, Nutrient.class);
@@ -72,8 +84,16 @@ public class NutrientServiceImpl implements NutrientService {
     }
 
     @Override
-    public void deleteNutrient(Long id) {
+    public void deleteNutrient(Long id) throws NutrientUsedInIngredient, NutrientNotFoundException {
         Nutrient existing = checkIfNutrientExistsThenGet(id);
+
+        boolean isUsed = ingredientRepository.existsByNutrient(existing);
+
+        if (isUsed) {
+            String name = existing.getName();
+            throw new NutrientUsedInIngredient(name);
+        }
+
         repository.delete(existing);
     }
 
@@ -83,7 +103,7 @@ public class NutrientServiceImpl implements NutrientService {
             .collect(Collectors.toList());
     }
 
-    private Nutrient checkIfNutrientExistsThenGet(Long id) {
+    private Nutrient checkIfNutrientExistsThenGet(Long id) throws NutrientNotFoundException {
         if (id != null) {
             Optional<Nutrient> found = repository.findById(id);
 
@@ -91,7 +111,7 @@ public class NutrientServiceImpl implements NutrientService {
                 return found.get();
             }
         }
-        throw new NotFoundException(idNotFound);
+        throw new NutrientNotFoundException(idNotFound);
     }
 
     private void checkIfNameExists(String nutrientName) throws NutrientExistsException {
